Fixed an overflow issues where functions with many terms overflowed the array size.

+++ psl-core/src/main/java/org/linqs/psl/model/rule/arithmetic/UnweightedGroundArithmeticRule.java
@@ -75,7 +75,7 @@ public class UnweightedGroundArithmeticRule extends AbstractGroundArithmeticRule
 
     @Override
     public ConstraintTerm getConstraintDefinition() {
-        GeneralFunction sum = new GeneralFunction(false, false, (short)coefficients.length);
+        GeneralFunction sum = new GeneralFunction(false, false, coefficients.length);
         for (int i = 0; i < coefficients.length; i++) {
             // Skip any grounding only predicates.
             if (atoms[i].getPredicate() instanceof GroundingOnlyPredicate) {
